resolve symlinks in content helpers
